Create the SFML with the entity and event system so now we can print the window with the backgound who move, and the sprite of the player who is an entity. Now gonna catch the event and make the player move

+++ source/ecs/Entity/Server/Fire.hpp
@@ -12,7 +12,7 @@ class Fire : public IMovingEntity
 {
 private:
     std::vector<std::shared_ptr<IComponent>> _component;
-    size_t _id;
+    int _id;
     bool _isPlayer;
 
 public:
